simulator/requirements.txt
mysql-connector-python
paho-mqtt==2.1.0
python-dotenv
pandas
openpyxl